fix xlnet
