+++ baselines/models/xlnet/run_classifier_lcqmc.sh
@@ -1,6 +1,6 @@
 export CUDA_VISIBLE_DEVICES="1"
-CURRENT_DIR=`pwd`
-
+pip install sentencepiece --user
+CURRENT_DIR=$(cd -P -- "$(dirname -- "$0")" && pwd -P)
 XLNET_DIR=$CURRENT_DIR/prev_trained_model/chinese_xlnet_mid_L-24_H-768_A-12
 OUTPUT_DIR=$CURRENT_DIR/lcqmc_output
 RAW_DIR=$CURRENT_DIR/../../glue/chineseGLUEdatasets/
